+++ src/context/ToastContextProvider.tsx
@@ -1,11 +1,11 @@
-import React, { createContext, useCallback, useContext, useState } from 'react';
+import React, { createContext, useCallback, useState } from 'react';
 import ToastMessage from '../components/common/ToastMessage';
 
 type ToastContextType = {
   showToast: (message: string) => void;
 };
 
-const ToastContext = createContext<ToastContextType | null>(null);
+export const ToastContext = createContext<ToastContextType | null>(null);
 
 export const ToastProvider = ({ children }: { children: React.ReactNode }) => {
   const [toast, setToast] = useState<string | null>(null);
@@ -20,12 +20,4 @@ export const ToastProvider = ({ children }: { children: React.ReactNode }) => {
       {toast && <ToastMessage message={toast} onClose={() => setToast(null)} />}
     </ToastContext.Provider>
   );
-};
-
-export const useToast = () => {
-  const context = useContext(ToastContext);
-  if (!context) {
-    throw new Error('useToast must be used within a ToastProvider');
-  }
-  return context;
 };
