+++ src/hooks/useToast.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ToastContext } from '../context/ToastContextProvider';
+
+export const useToast = () => {
+  const context = useContext(ToastContext);
+  if (!context) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+  return context;
+};
